web/requirements.txt
Flask-SQLAlchemy==1.0
Flask-WTF==0.9.3
Flask-login==0.3.1
backports.pbkdf2==0.1